fix grammar problems
This fills in a missing word (with what I _guessed_ should be there) and a couple other apparent fixes.  (It always possible that I misunderstood the intended meaning and my "fix" was actually worse, like a bad Siri autocorrection.)

Since the three changes were all on the same line, and submitting three different changes would create conflicts upon merging them, I just made the three changes in one commit.  So if any of them need to be rejected, it'll require a manual change.